+++ packages/common/src/types/scheduler.ts
@@ -1,3 +1,4 @@
+import type { CreateProject } from '..';
 import assertUnreachable from '../utils/assertUnreachable';
 import { type Explore, type ExploreError } from './explore';
 import { type DashboardFilterRule } from './filter';
@@ -46,6 +47,7 @@ export type SchedulerLog = {
         | 'uploadGsheets'
         | 'downloadCsv'
         | 'uploadGsheetFromQuery'
+        | 'createProjectWithoutCompile'
         | 'compileProject'
         | 'testAndCompileProject'
         | 'validateProject'
@@ -314,6 +316,13 @@ export type ApiTestSchedulerResponse = {
     };
 };
 
+export type ApiSchedulerJobIdResponse = {
+    status: 'ok';
+    results: {
+        jobId: string;
+    };
+};
+
 // Scheduler task types
 export type ScheduledDeliveryPayload =
     | { schedulerUuid: string }
@@ -405,6 +414,14 @@ export type ApiCsvUrlResponse = {
         status: string;
         truncated: boolean;
     };
+};
+
+export type SchedulerCreateProjectWithoutCompilePayload = {
+    createdByUserUuid: string;
+    organizationUuid: string;
+    requestMethod: string;
+    isPreview: boolean;
+    data: CreateProject;
 };
 
 export type CompileProjectPayload = {